Added configuration saving/loading.
- Fixed reading of wrong sized behaviors.
- Modified behavior speed to support fractional speeds.
- Configuration window can now be hidden/shown  by doubleclicking on trary icon.
- Added styling of configuration window.
- Beautified toolbar in configuration window.
- Added support for mouseover behaviors.
- Fixed bug with segfault when ponies had no speech lines.

+++ behavior.h
@@ -36,7 +36,7 @@ public:
         Dragged			= 1 << 6
     };
 
-    uint32_t speed;
+    float speed;
     int width;
     int height;
     QMovie* current_animation;
